test(libterrain,libindex): split datasets and base-filters by behaviour family
- datasets.test.js (433) → -conditions (datasets node condition
  resolution) and -fhir (fhir-cross-ref + fhir-microdata-html nodes);
  each block's local helpers travel with its describe.
- base-filters.test.js (454) → -query (Shared Filter Logic + New
  IndexBase Implementation) and -contract (Abstract Method Enforcement +
  Error Handling and Edge Cases); shared TestIndex lifted to -helpers.js.

Assertions unchanged; test counts preserved (14, 19).
Spec 0640 § C (large files), design Decision 5.
